Fix compiler errors and warnings

+++ src/Animation.cpp
@@ -7,11 +7,16 @@ Animation::Animation(const std::string& name, const sf::Texture& texture)
 Animation::Animation(const std::string& name, const sf::Texture& texture, int frame_count, int speed)
 		: m_name(name), m_sprite(texture), m_frame_count(frame_count), m_current_frame(0), m_speed(speed) {
 		m_size = Vec2(
-			static_cast<float>(texture.getSize().x / frame_count),
-			static_cast<float>(texture.getSize().y / frame_count)
+			static_cast<float>(texture.getSize().x / static_cast<unsigned int>(frame_count)),
+			static_cast<float>(texture.getSize().y / static_cast<unsigned int>(frame_count))
 		);
 		m_sprite.setOrigin(m_size.x / 2.0f, m_size.y / 2.0f);
-		m_sprite.setTextureRect(sf::IntRect(std::floor(m_current_frame) * m_size.x, 0, m_size.x, m_size.y));
+		m_sprite.setTextureRect(sf::IntRect(
+			static_cast<int>(std::floor(m_current_frame)) * static_cast<int>(m_size.x),
+			0,
+			static_cast<int>(m_size.x),
+			static_cast<int>(m_size.y)
+		));
 }
 
 void Animation::update() {
